move ee repositories endpoints to v3
Issue: AAH-766
