+++ CHANGES/766.feature
@@ -1,0 +1,1 @@
+Move execution-environment repository endpoints to v3 viewsets and remove execution-environment/namespaces
