+++ galaxy_ng/app/api/ui/serializers/__init__.py
@@ -28,15 +28,7 @@ from .distribution import (
 )
 
 from .execution_environment import (
-    ContainerRepositorySerializer,
-    ContainerManifestSerializer,
-    ContainerRepositoryHistorySerializer,
-    ContainerReadmeSerializer,
-    ContainerManifestDetailSerializer,
-    ContainerNamespaceDetailSerializer,
-    ContainerRegistryRemoteSerializer,
-    ContainerTagSerializer,
-    ContainerRemoteSerializer,
+    ContainerRegistryRemoteSerializer
 )
 
 __all__ = (
@@ -63,12 +55,6 @@ __all__ = (
     # container
     'ContainerRepositorySerializer',
     'ContainerRepositoryImageSerializer',
-    'ContainerRepositoryHistorySerializer',
-    'ContainerManifestSerializer',
-    'ContainerManifestDetailSerializer',
-    'ContainerReadmeSerializer',
-    'ContainerNamespaceDetailSerializer',
     'ContainerRegistryRemoteSerializer',
-    'ContainerRemoteSerializer',
-    'ContainerTagSerializer'
+    'ContainerRemoteSerializer'
 )
