+++ galaxy_ng/app/api/ui/serializers/execution_environment.py
@@ -1,22 +1,17 @@
-import json
 import re
 
 from rest_framework import serializers
+from drf_spectacular.utils import extend_schema_field
 from django.core import exceptions
 from django.db import transaction
 from django.utils.translation import gettext_lazy as _
-from drf_spectacular.types import OpenApiTypes
-from drf_spectacular.utils import extend_schema_field
 
-from pulpcore.plugin.util import get_users_with_perms
+from pulpcore.plugin import serializers as core_serializers
 
 from pulp_container.app import models as container_models
 from pulp_container.app import serializers as container_serializers
-from pulpcore.plugin import models as core_models
-from pulpcore.plugin import serializers as core_serializers
 
 from galaxy_ng.app import models
-from galaxy_ng.app.access_control.fields import GroupPermissionField, MyPermissionsField
 from galaxy_ng.app.api import utils
 
 namespace_fields = ("name", "my_permissions", "owners")
@@ -24,273 +19,10 @@ namespace_fields = ("name", "my_permissions", "owners")
 VALID_REMOTE_REGEX = r"^[A-Za-z0-9._-]*/?[A-Za-z0-9._-]*$"
 
 
-class ContainerNamespaceSerializer(serializers.ModelSerializer):
-    my_permissions = MyPermissionsField(source="*", read_only=True)
-    owners = serializers.SerializerMethodField()
-
-    class Meta:
-        model = models.ContainerNamespace
-        fields = namespace_fields
-        read_only_fields = (
-            "name",
-            "my_permissions",
-        )
-
-    @extend_schema_field(serializers.ListField)
-    def get_owners(self, namespace):
-        return get_users_with_perms(
-            namespace, with_group_users=False, for_concrete_model=True
-        ).values_list("username", flat=True)
-
-
-class ContainerNamespaceDetailSerializer(ContainerNamespaceSerializer):
-    groups = GroupPermissionField()
-
-    class Meta:
-        model = models.ContainerNamespace
-        fields = namespace_fields + ("groups",)
-        read_only_fields = (
-            "name",
-            "my_permissions",
-        )
-
-
-class ContainerRepositorySerializer(serializers.ModelSerializer):
-    pulp = serializers.SerializerMethodField()
-    namespace = ContainerNamespaceSerializer()
-    id = serializers.SerializerMethodField()
-    created = serializers.SerializerMethodField()
-    updated = serializers.SerializerMethodField()
-
-    # This serializer is purposfully refraining from using pulp fields directly
-    # in the top level response body. This is because future versions of hub will have to
-    # support indexing other registries and the API responses for a container
-    # repository should make sense for containers hosted by pulp and containers
-    # hosted by other registries.
-    class Meta:
-        model = models.ContainerDistribution
-        read_only_fields = (
-            "id",
-            "name",
-            # this field will return null on instances where hub is indexing a
-            # different repo
-            "pulp",
-            "namespace",
-            "description",
-            "created",
-            "updated",
-        )
-
-        fields = read_only_fields
-
-    def get_namespace(self, distro) -> str:
-        return distro.namespace.name
-
-    @extend_schema_field(OpenApiTypes.UUID)
-    def get_id(self, distro):
-        return distro.pk
-
-    @extend_schema_field(OpenApiTypes.DATETIME)
-    def get_created(self, distro):
-        return distro.repository.pulp_created
-
-    @extend_schema_field(OpenApiTypes.DATETIME)
-    def get_updated(self, distro):
-        return distro.repository.pulp_last_updated
-
-    @extend_schema_field(OpenApiTypes.OBJECT)
-    def get_pulp(self, distro):
-        repo = distro.repository
-        remote = None
-        if repo.remote:
-            remote = ContainerRemoteSerializer(repo.remote.cast()).data
-
-        sign_state = repo.content.filter(
-            pulp_type="container.signature"
-        ).count() > 0 and "signed" or "unsigned"
-
-        return {
-            "repository": {
-                "pulp_id": repo.pk,
-                "pulp_type": repo.pulp_type,
-                "version": repo.latest_version().number,
-                "name": repo.name,
-                "description": repo.description,
-                "pulp_created": repo.pulp_created,
-                "pulp_labels": {label.key: label.value for label in repo.pulp_labels.all()},
-                "remote": remote,
-                "sign_state": sign_state
-            },
-            "distribution": {
-                "pulp_id": distro.pk,
-                "name": distro.name,
-                "pulp_created": distro.pulp_created,
-                "base_path": distro.base_path,
-                "pulp_labels": {label.key: label.value for label in distro.pulp_labels.all()},
-            },
-        }
-
-
-class ManifestListManifestSerializer(serializers.ModelSerializer):
-    digest = serializers.SerializerMethodField()
-
-    class Meta:
-        model = container_models.ManifestListManifest
-        fields = (
-            "os",
-            "architecture",
-            "os_version",
-            "os_features",
-            "features",
-            "variant",
-            "digest"
-        )
-
-    def get_digest(self, obj):
-        return obj.manifest_list.digest
-
-
-class ContainerManifestSerializer(serializers.ModelSerializer):
-    config_blob = serializers.SerializerMethodField()
-    tags = serializers.SerializerMethodField()
-    layers = serializers.SerializerMethodField()
-    image_manifests = ManifestListManifestSerializer(many=True)
-
-    class Meta:
-        model = container_models.Manifest
-        fields = (
-            "pulp_id",
-            "digest",
-            "schema_version",
-            "media_type",
-            "config_blob",
-            "tags",
-            "pulp_created",
-            "layers",
-            "image_manifests"
-        )
-
-    @extend_schema_field(OpenApiTypes.OBJECT)
-    def get_layers(self, obj):
-        layers = []
-        # use the prefetched blob_list and artifact_list instead of obj.blobs and
-        # blob._artifacts to cut down on queries made.
-        for blob in obj.blob_list:
-            # Manifest can be empty after deleting an image.
-            if blob.artifact_list:
-                layers.append(
-                    {"digest": blob.digest, "size": blob.artifact_list[0].size}
-                )
-        return layers
-
-    @extend_schema_field(OpenApiTypes.OBJECT)
-    def get_config_blob(self, obj):
-        if not obj.config_blob:
-            return {}
-        return {"digest": obj.config_blob.digest}
-
-    @extend_schema_field(serializers.ListField)
-    def get_tags(self, obj):
-        tags = []
-        # tagget_manifests returns all tags on the manifest, not just the ones
-        # that are in the latest version of the repo.
-        for tag in obj.tagged_manifests.all():
-            tags.append(tag.name)
-
-        return tags
-
-
-class ManifestListSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = container_models.Manifest
-        fields = (
-            "pulp_id",
-            "digest",
-            "schema_version",
-            "media_type",
-            "pulp_created",
-        )
-
-
-class ContainerTagSerializer(serializers.ModelSerializer):
-    tagged_manifest = ManifestListSerializer()
-
-    class Meta:
-        model = container_models.Tag
-        fields = (
-            "name",
-            "pulp_created",
-            "pulp_last_updated",
-            "tagged_manifest"
-        )
-
-
-class ContainerManifestDetailSerializer(ContainerManifestSerializer):
-    @extend_schema_field(OpenApiTypes.OBJECT)
-    def get_config_blob(self, obj):
-        with obj.config_blob._artifacts.first().file.open() as f:
-            config_json = json.load(f)
-
-        return {
-            "digest": obj.config_blob.digest,
-            "data": config_json,
-        }
-
-
-class ContainerRepositoryHistorySerializer(serializers.ModelSerializer):
-    added = serializers.SerializerMethodField()
-    removed = serializers.SerializerMethodField()
-
-    class Meta:
-        model = core_models.RepositoryVersion
-        fields = ("pulp_id", "added", "removed", "pulp_created", "number")
-
-    @extend_schema_field(serializers.ListField(child=serializers.JSONField()))
-    def get_added(self, obj):
-        return [self._content_info(content.content) for content in obj.added_memberships.all()]
-
-    @extend_schema_field(serializers.ListField(child=serializers.JSONField()))
-    def get_removed(self, obj):
-        return [self._content_info(content.content) for content in obj.removed_memberships.all()]
-
-    def _content_info(self, content):
-        return_data = {
-            "pulp_id": content.pk,
-            "pulp_type": content.pulp_type,
-            "manifest_digest": None,
-            "tag_name": None,
-        }
-
-        # TODO: Figure out if there is a way to prefetch Manifest and Tag objects
-        if content.pulp_type == "container.manifest":
-            manifest = container_models.Manifest.objects.get(pk=content.pk)
-            return_data["manifest_digest"] = manifest.digest
-        elif content.pulp_type == "container.tag":
-            tag = container_models.Tag.objects.select_related("tagged_manifest").get(pk=content.pk)
-            return_data["manifest_digest"] = tag.tagged_manifest.digest
-            return_data["tag_name"] = tag.name
-
-        return return_data
-
-
-class ContainerReadmeSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = models.ContainerDistroReadme
-        fields = (
-            "updated",
-            "created",
-            "text",
-        )
-
-        read_only_fields = (
-            "updated",
-            "created",
-        )
-
-
 class ContainerRegistryRemoteSerializer(
     core_serializers.RemoteSerializer,
 ):
+    id = serializers.UUIDField(source='pk', required=False)
     created_at = serializers.DateTimeField(source='pulp_created', required=False)
     updated_at = serializers.DateTimeField(source='pulp_last_updated', required=False)
     last_sync_task = utils.RemoteSyncTaskField(source="*")
@@ -300,7 +32,8 @@ class ContainerRegistryRemoteSerializer(
     class Meta:
         model = models.ContainerRegistryRemote
         fields = [
-            "pk",
+            "id",
+            "pulp_href",
             "name",
             "url",
             "policy",
@@ -339,6 +72,7 @@ class ContainerRegistryRemoteSerializer(
 class ContainerRemoteSerializer(
     container_serializers.ContainerRemoteSerializer,
 ):
+    id = serializers.UUIDField(source='pulp_id', required=False, read_only=True)
     created_at = serializers.DateTimeField(source='pulp_created', read_only=True, required=False)
     updated_at = serializers.DateTimeField(
         source='pulp_last_updated', read_only=True, required=False)
@@ -359,7 +93,8 @@ class ContainerRemoteSerializer(
         }
 
         fields = [
-            "pulp_id",
+            "id",
+            "pulp_href",
             "name",
             "upstream_name",
             "registry",
