+++ galaxy_ng/app/api/v3/serializers/__init__.py
@@ -16,6 +16,17 @@ from .task import (
     TaskSummarySerializer,
 )
 
+from .execution_environment import (
+    ContainerRepositorySerializer,
+    ContainerTagSerializer,
+    ContainerManifestSerializer,
+    ContainerManifestDetailSerializer,
+    ContainerReadmeSerializer,
+    ContainerRepositoryHistorySerializer,
+    ContainerNamespaceDetailSerializer,
+    ContainerRemoteSerializer
+)
+
 __all__ = (
     'CollectionUploadSerializer',
     'GroupSummarySerializer',
@@ -24,4 +35,12 @@ __all__ = (
     'TaskSerializer',
     'TaskSummarySerializer',
     'UnpaginatedCollectionVersionSerializer',
+    'ContainerRepositorySerializer',
+    'ContainerRepositoryHistorySerializer',
+    'ContainerManifestSerializer',
+    'ContainerTagSerializer',
+    'ContainerManifestDetailSerializer',
+    'ContainerReadmeSerializer',
+    'ContainerNamespaceDetailSerializer',
+    'ContainerRemoteSerializer'
 )
