+++ galaxy_ng/app/api/v3/views/__init__.py
@@ -2,6 +2,7 @@ from .auth import TokenView
 from .sync import SyncRemoteView
 from .excludes import ExcludesView
 from .not_found import NotFoundView
+from .sync import ContainerSyncRemoteView
 
 
-__all__ = ("TokenView", "SyncRemoteView", "ExcludesView", "NotFoundView")
+__all__ = ("TokenView", "SyncRemoteView", "ExcludesView", "NotFoundView", "ContainerSyncRemoteView")
