+++ galaxy_ng/app/viewsets.py
@@ -5,7 +5,7 @@ from galaxy_ng.app import models
 from galaxy_ng.app.access_control import access_policy
 from galaxy_ng.app.access_control.statements.roles import LOCKED_ROLES as GALAXY_LOCKED_ROLES
 from galaxy_ng.app.api.ui import serializers
-from galaxy_ng.app.api.v3.serializers import NamespaceSummarySerializer
+from galaxy_ng.app.api.v3.serializers import ContainerRepositorySerializer
 
 # This file is necesary to prevent the DRF web API browser from breaking on all of the
 # pulp/api/v3/repositories/ endpoints.
@@ -40,30 +40,6 @@ class ContainerDistributionViewSet(
     mixins.RetrieveModelMixin,
 ):
     queryset = models.ContainerDistribution.objects.all()
-    serializer_class = serializers.ContainerRepositorySerializer
+    serializer_class = ContainerRepositorySerializer
     permission_classes = [access_policy.ContainerRepositoryAccessPolicy]
     endpoint_name = "galaxy_ng/container-distribution-proxy"
-
-
-# added so that object permissions are viewable for namespaces on the roles api endpoint.
-class NamespaceViewSet(
-    pulp_viewsets.NamedModelViewSet,
-    mixins.ListModelMixin,
-    mixins.RetrieveModelMixin,
-):
-    queryset = models.Namespace.objects.all()
-    serializer_class = NamespaceSummarySerializer
-    permission_classes = [access_policy.NamespaceAccessPolicy]
-    endpoint_name = "pulp_ansible/namespaces"
-
-
-# This is here because when new objects are created, pulp tries to look up the viewset for
-# the model so that it can run creation hooks to assign permissions. This function fails
-# if the model isn't registered with a pulp viewset
-# https://github.com/pulp/pulpcore/blob/52c8e16997849b9a639aec04945cec98486d54fb/pulpcore
-#   /app/models/access_policy.py#L73
-class GroupViewset(
-    pulp_viewsets.NamedModelViewSet,
-):
-    queryset = models.auth.Group.objects.all()
-    endpoint_name = "galaxy_ng/sgroups"
